+++ angular/projects/demo/src/app/app.config.ts
@@ -1,7 +1,5 @@
 import { ApplicationConfig, provideEnvironmentInitializer } from '@angular/core';
 
-// TEST local testing
-// import { GridstackComponent } from './gridstack.component';
 import { GridstackComponent } from 'gridstack/dist/angular';
 import { AComponent, BComponent, CComponent, NComponent } from './dummy.component';
 
@@ -9,7 +7,7 @@ export const appConfig: ApplicationConfig = {
   providers: [
     provideEnvironmentInitializer(() => {
       // register all our dynamic components created in the grid
-      GridstackComponent.addComponentToSelectorType([AComponent, BComponent, CComponent, NComponent]);
+      GridstackComponent.registerComponents([AComponent, BComponent, CComponent, NComponent]);
     })
   ]
 };